nit: docstrings
