+++ libs/azure-ai/langchain_azure_ai/agents/prebuilt/tools.py
@@ -1,6 +1,12 @@
 """Azure AI Foundry Agent Service Tools."""
 
-from azure.ai.agents.models import Tool
+from typing import Any, List
+
+from azure.ai.agents.models import (
+    FunctionToolDefinition,
+    Tool,
+    ToolResources,
+)
 from pydantic import BaseModel, ConfigDict
 
 
@@ -12,7 +18,7 @@ class AgentServiceBaseTool(BaseModel):
 
     Example:
         ... code-block:: python
-            from langchain_azure_ai.tools.agent_service import AgentServiceBaseTool
+            from langchain_azure_ai.agents.prebuilt.tools import AgentServiceBaseTool
             from azure.ai.agents.models import CodeInterpreterTool
 
             code_interpreter_tool = AgentServiceBaseTool(tool=CodeInterpreterTool())
@@ -23,3 +29,41 @@ class AgentServiceBaseTool(BaseModel):
 
     tool: Tool
     """The tool definition from Azure AI Foundry."""
+
+
+class _OpenAIFunctionTool(Tool[FunctionToolDefinition]):
+    """A tool that wraps OpenAI function definitions."""
+
+    def __init__(self, definitions: List[FunctionToolDefinition]):
+        """Initialize the OpenAIFunctionTool with function definitions.
+
+        Args:
+        definitions: A list of function definitions to be used by the tool.
+        """
+        self._definitions = definitions
+
+    @property
+    def definitions(self) -> List[FunctionToolDefinition]:
+        """Get the function definitions.
+
+        Returns:
+            A list of function definitions.
+        """
+        return self._definitions
+
+    @property
+    def resources(self) -> ToolResources:
+        """Get the tool resources for the agent.
+
+        Returns:
+            The tool resources.
+        """
+        return ToolResources()
+
+    def execute(self, tool_call: Any) -> Any:
+        """Execute the tool with the provided tool call.
+
+        :param Any tool_call: The tool call to execute.
+        :return: The output of the tool operations.
+        """
+        pass
